requirements.txt
PyPDF2==3.0.1
pdfminer.six==20231228
langchain==0.3.23
langchain-core==0.3.54
openai==1.75.0
python-dotenv==1.0.0
markdown==3.8
tqdm==4.66.5
requests==2.31.0
# 内置模型服务器所需依赖
torch>=2.6.0
transformers>=4.34.0
fastapi>=0.115.12
uvicorn>=0.34.2
pydantic>=2.11.3
sentencepiece>=0.2.0
accelerate>=0.23.0
